// src/services/Users/User.types.ts
export interface UserTypes {
    user_id: number;
    first_name: string;
    last_name: string;
    user_name: string;
    email: string;
    is_email_verified: boolean;
    phone_number: string;
    is_phone_verified: boolean;
    role: string;
    gender: string;
    date_of_birth: string;
    profile_image: string;
    status: string;
}